Number of rooms changed to int in dtos

+++ ApartmentFinder/AdvertisingService/AdvertisingService.WebAPI/Dtos/AdvertisementDto.cs
@@ -8,7 +8,7 @@ public record AdvertisementInfoDto
     public string? District { get; init; }
     public required string StreetName { get; init; }
     public string? StreetNumber { get; init; }
-    public required float NumberOfRooms { get; init; }
+    public required int NumberOfRooms { get; init; }
     public required float Size { get; init; }
     public required int MonthlyPrice { get; init; }
     public required  DateTime UploadDate { get; init; }
@@ -26,7 +26,7 @@ public record AdvertisementDetailsDto
     public required string StreetName { get; init; }
     public string? StreetNumber { get; init; }
     public string? UnitNumber { get; init; }
-    public required float NumberOfRooms { get; init; }
+    public required int NumberOfRooms { get; init; }
     public required float Size { get; init; }
     public required bool Furnished { get; init; }
     public required bool Parking { get; init; }
